Revamp the benchmark generation to account for fuel usage
This commit introduces a way to specify that fuel measurements are
desired, this is done via the `--fuel` flag. This will run the
specified benchmark and report execution time along with the number of
fuel units consumed.

This commit also revamps the statistical driven benchmarks to measure
compilation and execution with and without fuel. The following
command can be used to call criterion and run either a compilation or
execution benchmark with or without fuel:

`cargo run -- --name=<exec|compile>:<name>:<default|with-fuel>`

For example, if we'd like to statistically measure the execution time
of  the `shootout-fib2` benchmark with fuel, we should run:

`cargo run -- --name=exec:shootout-fib2:with-fuel`